Refactor: move MockRequest and MockResponse into mock_http module
