+++ web_monitoring_diff/server/server.py
@@ -6,7 +6,6 @@ import hashlib
 import inspect
 import functools
 import logging
-import mimetypes
 import os
 import pycurl
 import re
@@ -20,6 +19,7 @@ import tornado.ioloop
 import tornado.web
 import traceback
 import web_monitoring_diff
+from .mock_http import MockResponse
 from .. import basic_diffs, html_render_diff, html_links_diff
 from ..exceptions import UndiffableContentError, UndecodableContentError
 from ..utils import shutdown_executor_in_loop, Signal
@@ -167,43 +167,6 @@ class PublicError(tornado.web.HTTPError):
                 log_message = public_message
 
         super().__init__(status_code, log_message, **kwargs)
-
-
-class MockRequest:
-    "An HTTPRequest-like object for local file:/// requests."
-    def __init__(self, url):
-        self.url = url
-
-
-class MockResponse:
-    "An HTTPResponse-like object for local file:/// requests."
-    def __init__(self, url, body, headers=None):
-        self.request = MockRequest(url)
-        self.body = body
-        self.headers = headers
-        self.error = None
-
-        if self.headers is None:
-            self.headers = {}
-
-        if 'Content-Type' not in self.headers:
-            self.headers.update(self._get_content_type_headers_from_url(url))
-
-    @staticmethod
-    def _get_content_type_headers_from_url(url):
-        # If the extension is not recognized, assume text/html
-        headers = {'Content-Type': 'text/html'}
-
-        content_type, content_encoding = mimetypes.guess_type(url)
-
-        if content_type is not None:
-            headers['Content-Type'] = content_type
-
-        if content_encoding is not None:
-            headers['Content-Encoding'] = content_encoding
-
-        return headers
-
 
 DEBUG_MODE = os.environ.get('DIFFING_SERVER_DEBUG', 'False').strip().lower() == 'true'
